requirements.txt
langchain
langchain-openai
pydantic
python-dotenv
PyMuPDF
pillow
numpy
azure-ai-formrecognizer
fastapi
uvicorn
python-multipart